+++ src/v0_4/bioformats2raw_layout.rs
@@ -14,9 +14,7 @@ pub struct Bioformats2rawLayout {
 
 #[cfg(test)]
 mod tests {
-    use serde_json::Value;
-
-    use crate::v0_4::Ome;
+    use crate::v0_4::OmeNgffGroupAttributes;
 
     use super::*;
 
@@ -26,8 +24,7 @@ mod tests {
             env!("CARGO_MANIFEST_DIR"),
             "/ome-zarr/0.4/examples/bf2raw/image.json"
         ));
-        let ome_metadata: Value = serde_json::from_str(&json).unwrap();
-        let ome_metadata: Ome = serde_json::from_value(ome_metadata.clone()).unwrap();
+        let ome_metadata: OmeNgffGroupAttributes = serde_json::from_str(json).unwrap();
         let _bioformats2raw: Bioformats2rawLayout = ome_metadata.bioformats2raw_layout.unwrap();
     }
 
@@ -37,8 +34,7 @@ mod tests {
             env!("CARGO_MANIFEST_DIR"),
             "/ome-zarr/0.4/examples/bf2raw/plate.json"
         ));
-        let ome_metadata: Value = serde_json::from_str(&json).unwrap();
-        let ome_metadata: Ome = serde_json::from_value(ome_metadata.clone()).unwrap();
+        let ome_metadata: OmeNgffGroupAttributes = serde_json::from_str(json).unwrap();
         let _bioformats2raw: Bioformats2rawLayout = ome_metadata.bioformats2raw_layout.unwrap();
     }
 }