+++ src/v0_4/labels.rs
@@ -53,9 +53,7 @@ pub struct ImageLabelSource {
 
 #[cfg(test)]
 mod tests {
-    use serde_json::Value;
-
-    use crate::v0_4::Ome;
+    use crate::v0_4::OmeNgffGroupAttributes;
 
     use super::*;
 
@@ -65,8 +63,7 @@ mod tests {
             env!("CARGO_MANIFEST_DIR"),
             "/ome-zarr/0.4/examples/label_strict/colors_properties.json"
         ));
-        let ome_metadata: Value = serde_json::from_str(&json).unwrap();
-        let ome_metadata: Ome = serde_json::from_value(ome_metadata.clone()).unwrap();
+        let ome_metadata: OmeNgffGroupAttributes = serde_json::from_str(json).unwrap();
         let _image_label: ImageLabel = ome_metadata.image_label.unwrap();
     }
 }