+++ src/v0_4/multiscales.rs
@@ -48,9 +48,7 @@ pub struct MultiscaleImageMetadata(pub serde_json::Map<String, serde_json::Value
 
 #[cfg(test)]
 mod tests {
-    use serde_json::Value;
-
-    use crate::v0_4::Ome;
+    use crate::v0_4::OmeNgffGroupAttributes;
 
     use super::*;
 
@@ -63,9 +61,7 @@ mod tests {
         .lines()
         .filter(|line| !line.contains("//")) // Remove comments
         .collect::<String>();
-
-        let ome_metadata: Value = serde_json::from_str(&json).unwrap();
-        let ome_metadata: Ome = serde_json::from_value(ome_metadata.clone()).unwrap();
+        let ome_metadata: OmeNgffGroupAttributes = serde_json::from_str(&json).unwrap();
         let _multiscales: Vec<MultiscaleImage> = ome_metadata.multiscales.unwrap();
     }
 
@@ -75,8 +71,7 @@ mod tests {
             env!("CARGO_MANIFEST_DIR"),
             "/ome-zarr/0.4/examples/multiscales_strict/multiscales_transformations.json"
         ));
-        let ome_metadata: Value = serde_json::from_str(&json).unwrap();
-        let ome_metadata: Ome = serde_json::from_value(ome_metadata.clone()).unwrap();
+        let ome_metadata: OmeNgffGroupAttributes = serde_json::from_str(json).unwrap();
         let _multiscales: Vec<MultiscaleImage> = ome_metadata.multiscales.unwrap();
     }
 }