+++ src/v0_4/plate.rs
@@ -84,9 +84,7 @@ pub struct PlateWell {
 
 #[cfg(test)]
 mod tests {
-    use serde_json::Value;
-
-    use crate::v0_4::Ome;
+    use crate::v0_4::OmeNgffGroupAttributes;
 
     use super::*;
 
@@ -96,8 +94,7 @@ mod tests {
             env!("CARGO_MANIFEST_DIR"),
             "/ome-zarr/0.4/examples/plate_strict/plate_2wells.json"
         ));
-        let ome_metadata: Value = serde_json::from_str(&json).unwrap();
-        let ome_metadata: Ome = serde_json::from_value(ome_metadata.clone()).unwrap();
+        let ome_metadata: OmeNgffGroupAttributes = serde_json::from_str(json).unwrap();
         let _plate: Plate = ome_metadata.plate.unwrap();
     }
 
@@ -107,8 +104,7 @@ mod tests {
             env!("CARGO_MANIFEST_DIR"),
             "/ome-zarr/0.4/examples/plate_strict/plate_6wells.json"
         ));
-        let ome_metadata: Value = serde_json::from_str(&json).unwrap();
-        let ome_metadata: Ome = serde_json::from_value(ome_metadata.clone()).unwrap();
+        let ome_metadata: OmeNgffGroupAttributes = serde_json::from_str(json).unwrap();
         let _plate: Plate = ome_metadata.plate.unwrap();
     }
 }