+++ src/v0_5/plate.rs
@@ -31,7 +31,7 @@ pub struct Plate {
 
 #[cfg(test)]
 mod tests {
-    use crate::v0_5::{get_ome_attribute_from_zarr_group_metadata, Ome};
+    use crate::v0_5::OmeZarrGroupMetadata;
 
     use super::*;
 
@@ -41,11 +41,8 @@ mod tests {
             env!("CARGO_MANIFEST_DIR"),
             "/ome-zarr/0.5/examples/plate_strict/plate_2wells.json"
         ));
-        let group_metadata: serde_json::Map<String, serde_json::Value> =
-            serde_json::from_str(json).unwrap();
-        let ome_metadata = get_ome_attribute_from_zarr_group_metadata(&group_metadata).unwrap();
-        let ome_metadata: Ome = serde_json::from_value(ome_metadata.clone()).unwrap();
-        let _plate: Plate = ome_metadata.plate.unwrap();
+        let ome_metadata: OmeZarrGroupMetadata = serde_json::from_str(json).unwrap();
+        let _plate: Plate = ome_metadata.attributes.ome.plate.unwrap();
     }
 
     #[test]
@@ -54,10 +51,7 @@ mod tests {
             env!("CARGO_MANIFEST_DIR"),
             "/ome-zarr/0.5/examples/plate_strict/plate_6wells.json"
         ));
-        let group_metadata: serde_json::Map<String, serde_json::Value> =
-            serde_json::from_str(json).unwrap();
-        let ome_metadata = get_ome_attribute_from_zarr_group_metadata(&group_metadata).unwrap();
-        let ome_metadata: Ome = serde_json::from_value(ome_metadata.clone()).unwrap();
-        let _plate: Plate = ome_metadata.plate.unwrap();
+        let ome_metadata: OmeZarrGroupMetadata = serde_json::from_str(json).unwrap();
+        let _plate: Plate = ome_metadata.attributes.ome.plate.unwrap();
     }
 }